refactor: move all v2 routes to default
