+++ src/app/page.tsx
@@ -1,5 +1,15 @@
+import HomeScreen from "@/screens/home";
+import { auth } from "@clerk/nextjs/server";
 import { redirect } from "next/navigation";
 
-export default function RootPage() {
-  redirect("/v2");
-}
+const HomePage = () => {
+  const { userId } = auth();
+
+  if (userId) {
+    return redirect("/chat");
+  }
+
+  return <HomeScreen />;
+};
+
+export default HomePage;
