+++ src/middleware.ts
@@ -1,11 +1,6 @@
 import { clerkMiddleware, createRouteMatcher } from "@clerk/nextjs/server";
 
-const isProtectedRoute = createRouteMatcher([
-  "/v2/chat(.*)",
-  "/v2/settings(.*)",
-  "/v1/chat-compare(.*)",
-  "/v1/account(.*)",
-]);
+const isProtectedRoute = createRouteMatcher(["/chat(.*)", "/settings(.*)"]);
 
 export default clerkMiddleware((auth, req) => {
   if (isProtectedRoute(req)) auth().protect();
